% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_mirror}
\alias{env_mirror}
\title{Mirror an environment}
\usage{
env_mirror(env, parent = env_parent(env))
}
\arguments{
\item{env}{An environment to clone.}

\item{parent}{The parent of the new mirror environment.}
}
\description{
A mirror is an environment that shares the same objects but not the
same parent. It is an efficient way of cloning an environment when
the purpose of the cloning is to have a different parent.
}
\examples{
env <- env()
mirror <- env_mirror(env)

# A mirror is like a shallow clone. The contents of the mirror are
# shared with the original environment:
mirror$a <- "foo"
env$a

# On the other hand, the mirror can have a different parent:
mut_env_parent(mirror, base_env())
env_parent(mirror)
env_parent(env)
}
